pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "knowme"
version = "1.0.0"
authors = [
  { name="Your Name", email="you@example.com" },
]
description = "A fast, offline command-line tool that displays detailed system info with a classic, neofetch-style ASCII logo."
readme = "README.md"
requires-python = ">=3.7"
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
]
dependencies = [
    "psutil",
    "distro",
    "py-cpuinfo",
    "requests",
    "gputil",
    "screeninfo",
    "ifaddr",
]

[project.urls]
"Homepage" = "https://github.com/pypa/sampleproject"
"Bug Tracker" = "https://github.com/pypa/sampleproject/issues"

[project.scripts]
knowme = "knowme.__main__:main"